+++ NEWS.md
@@ -1,3 +1,20 @@
+# twn 0.2.0
+
+## Nieuw
+
+- `twn_taxontype()` toegevoegd om het taxontype op te vragen. (NB sommige taxa hebben meer dan 1 taxontype)
+- `twn_children()` functie om alle child-taxa van een bepaalde parent te vinden.
+- `is_taxonlevel()`, `is_taxontype()` en `is_status` toegevoegd. Functies om te checken of taxa een bepaald(e) taxonlevel, taxontype of status hebben.
+
+## Wijzigingen
+
+- Nieuwe TWN-lijst (17-3-2021) 43 wijzigingen t.o.v. de vorige versie.
+
+## TODO
+
+- `twn_synonyms`
+- `has_parent` `has_child`
+
 # twn 0.1.1